Vc 2606 helpers to typescript (#2118)
* VC-2606 Rewrite helpers component to TypeScript

* VC-2606 Fix TS and codestyle issues
